# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.RawHop:
    properties:
      hash:
        description: hex-encoded path hash
        type: string
      snr:
        description: signal-to-noise ratio in dB, if available
        type: number
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop:
    properties:
      confidence:
        description: '"high", "ambiguous", or "none"'
        type: string
      nodes:
        description: empty for "none", one for "high", multiple for "ambiguous"
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode'
        type: array
      snr:
        type: number
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode:
    properties:
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      publicKey:
        description: hex-encoded prefix used for resolution
        type: string
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.TraceDetail:
    properties:
      packets:
        description: all packets observed for this trace
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.TracePacket'
        type: array
      traceTag:
        description: hex-encoded 4-byte tag
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.TracePacket:
    properties:
      firstHeardAt:
        description: epoch ms
        type: integer
      lastHeardAt:
        description: epoch ms
        type: integer
      packetHash:
        description: hex-encoded packet hash
        type: string
      rawPath:
        description: hops as received in the packet
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.RawHop'
        type: array
      resolvedRoute:
        description: hops resolved to known nodes
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop'
        type: array
      routeType:
        description: numeric route type
        type: integer
      routeTypeName:
        description: human-readable route type
        type: string
      scope:
        description: transport scope name, if known
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.TraceTagSummary:
    properties:
      firstHeardAt:
        description: epoch ms
        type: integer
      iataCount:
        description: number of distinct IATAs where heard
        type: integer
      lastHeardAt:
        description: epoch ms
        type: integer
      packetCount:
        description: number of packets with this trace tag
        type: integer
      pathHashes:
        description: hops from the most complete observation
        items:
          type: string
        type: array
      snrValues:
        description: SNR per hop from the most complete observation
        items:
          type: number
        type: array
      traceTag:
        description: hex-encoded 4-byte tag
        type: string
      traceType:
        description: TRACE or PING
        type: string
    type: object
  internal_api_handlers.APIError:
    properties:
      code:
        description: e.g. "not_found", "bad_request"
        type: string
      message:
        description: e.g. "channel not found"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: MeshCore Beacon
    url: https://github.com/MeshCore-Beacon/beacon-server
  description: MeshCore network observation backend. Ingests LoRa packets from MQTT
    brokers, stores in PostgreSQL, and streams live events via WebSocket.
  license:
    name: AGPL-3-or-later
  termsOfService: https://github.com/MeshCore-Beacon/beacon-server
  title: MeshCore Beacon API
  version: 1.6.0
paths:
...
  /traces:
    get:
      parameters:
      - description: Filter by IATA code(s), comma-separated
        in: query
        name: iatas
        type: string
      - description: Filter by region slug
        in: query
        name: region
        type: string
      - description: Filter by region ID
        in: query
        name: regionId
        type: integer
      - description: Filter by transport scope name
        in: query
        name: scope
        type: string
      - description: 'Filter by type: TRACE or PING (default: all)'
        in: query
        name: type
        type: string
      - description: Filter by first_heard_at >= since (epoch ms)
        in: query
        name: since
        type: integer
      - description: Filter by first_heard_at <= until (epoch ms)
        in: query
        name: until
        type: integer
      - description: last_heard_at epoch ms of last item for pagination
        in: query
        name: cursor
        type: integer
      - description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.TraceTagSummary'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List trace tags
      tags:
      - Traces
  /traces/{tag}:
    get:
      parameters:
      - description: Trace tag hex e.g. a3f1b2c4
        in: path
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.TraceDetail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Get full trace detail by tag
      tags:
      - Traces
schemes:
- http
- https
swagger: "2.0"
...